requirements.txt
Pillow>=11.2.1
matplotlib>=3.10.1
aiogram>=3.20.0